requirements.txt
google-generativeai
python-dotenv
streamlit
tweepy
mistralai